+++ src/unitTests/als.cxx
@@ -128,13 +128,12 @@ UNIT_TEST(ALS, tutorial,
 	C(i&0) = A(i/2, j/2) * B(j&0);
 	X = xerus::TTTensor::random(stateDims, 2, rnd, dist);
 	
-	xerus::ALSVariant ALSb(xerus::ASD_SPD);
-	PerformanceData pd(true);
+	PerformanceData pd(false);
 // 	ALSb.printProgress = true;
 // 	ALSb.useResidualForEndCriterion = true;
 // 	std::vector<value_t> perfdata;
 	
-	ALSb(A, X, C, 1e-12, pd);
+	ALS_SPD(A, X, C, 1e-12, pd);
 	TEST(frob_norm(A(i/2, j/2)*X(j&0) - C(i&0)) < 1e-4);